+++ Documents/Help/Changelog.md
@@ -7,6 +7,7 @@
 * Full/Lite NuGet packages - as support for .NET 5 ended in May, there are now 2 types of NuGet package.
     - Full - Supports every framework from .NET Framework 4.6.2 to .NET 6
     - Lite - Supports .NET Framework 4.8, .NET Core 3.1 and .NET 6
+* Fixed grid cell selection colours for dark/light mode themes
 
 ## 2022-04-04 - Build 2204 - April 2022
 * Resolved [#678](https://github.com/Krypton-Suite/Standard-Toolkit/issues/678), Dropdown list background & text colour are the same (Office 2010 - Black (Dark Mode))